requirements.txt
pytest==8.4.2
selenium==4.40.0
allure-pytest==2.36.0